pyproject.toml
[project]
name = "hotcrowdcms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-allauth>=65.13.1",
    "django-htmx>=1.27.0",
    "django-ninja>=1.5.1",
    "gunicorn>=23.0.0",
    "instagrapi>=2.2.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "whitenoise>=6.11.0",
]
